refactor: 탭 바 아이콘 SVG로 교체
