+++ src/components/common/TabBar.jsx
@@ -1,0 +1,117 @@
+import styled from 'styled-components';
+
+import { NavLink } from 'react-router-dom';
+
+const Container = styled.div`
+  position: relative;
+
+  width: 24.375rem;
+  height: 3.8125rem;
+
+  display: flex;
+  justify-content: center;
+  gap: 3.56rem;
+
+  background-color: white;
+
+  .inactive {
+    opacity: 0.3;
+  }
+  .active {
+    opacity: 1;
+  }
+  /* transition: 1s ease all; */
+`;
+
+const IconContainer = styled.div`
+  width: 2rem;
+  height: 100%;
+
+  display: flex;
+  align-items: center;
+`;
+
+const TabBar = () => {
+  return (
+    <Container>
+      <NavLink
+        to={`/home`}
+        className={({ isActive }) => (isActive ? 'active' : 'inactive')}
+      >
+        <IconContainer>
+          <svg
+            width="32"
+            height="32"
+            viewBox="0 0 32 32"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              d="M7.99984 25.3332H12.9229V17.487H19.0768V25.3332H23.9998V13.3332L15.9998 7.28193L7.99984 13.3332V25.3332ZM6.6665 26.6665V12.6665L15.9998 5.61523L25.3332 12.6665V26.6665H17.7434V18.8204H14.2562V26.6665H6.6665Z"
+              fill="#1F1F1F"
+            />
+          </svg>
+        </IconContainer>
+      </NavLink>
+      <NavLink
+        to={`/collage`}
+        className={({ isActive }) => (isActive ? 'active' : 'inactive')}
+      >
+        <IconContainer>
+          <svg
+            width="32"
+            height="32"
+            viewBox="0 0 32 32"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              d="M7.48733 26.6663C6.87366 26.6663 6.36127 26.4608 5.95016 26.0497C5.53905 25.6386 5.3335 25.1262 5.3335 24.5125V7.48684C5.3335 6.87317 5.53905 6.36079 5.95016 5.94967C6.36127 5.53856 6.87366 5.33301 7.48733 5.33301H24.513C25.1267 5.33301 25.6391 5.53856 26.0502 5.94967C26.4613 6.36079 26.6668 6.87317 26.6668 7.48684V24.5125C26.6668 25.1262 26.4613 25.6386 26.0502 26.0497C25.6391 26.4608 25.1267 26.6663 24.513 26.6663H7.48733ZM7.48733 25.333H11.9976V20.0022H6.66683V24.5125C6.66683 24.7518 6.74375 24.9484 6.8976 25.1022C7.05144 25.2561 7.24802 25.333 7.48733 25.333ZM13.3309 25.333H18.6694V20.0022H13.3309V25.333ZM20.0027 25.333H24.513C24.7523 25.333 24.9489 25.2561 25.1027 25.1022C25.2566 24.9484 25.3335 24.7518 25.3335 24.5125V20.0022H20.0027V25.333ZM6.66683 18.6689H11.9976V13.3304H6.66683V18.6689ZM13.3309 18.6689H18.6694V13.3304H13.3309V18.6689ZM20.0027 18.6689H25.3335V13.3304H20.0027V18.6689ZM6.66683 11.9971H11.9976V6.66634H7.48733C7.24802 6.66634 7.05144 6.74326 6.8976 6.89711C6.74375 7.05095 6.66683 7.24753 6.66683 7.48684V11.9971ZM13.3309 11.9971H18.6694V6.66634H13.3309V11.9971ZM20.0027 11.9971H25.3335V7.48684C25.3335 7.24753 25.2566 7.05095 25.1027 6.89711C24.9489 6.74326 24.7523 6.66634 24.513 6.66634H20.0027V11.9971Z"
+              fill="#1F1F1F"
+            />
+          </svg>
+        </IconContainer>
+      </NavLink>
+      <NavLink
+        to={`/feed`}
+        className={({ isActive }) => (isActive ? 'active' : 'inactive')}
+      >
+        <IconContainer>
+          <svg
+            width="32"
+            height="32"
+            viewBox="0 0 32 32"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              d="M7.48733 26.6663C6.89075 26.6663 6.38263 26.4565 5.96296 26.0369C5.54332 25.6172 5.3335 25.1091 5.3335 24.5125V7.48684C5.3335 6.89026 5.54332 6.38214 5.96296 5.96247C6.38263 5.54283 6.89075 5.33301 7.48733 5.33301H24.513C25.1096 5.33301 25.6177 5.54283 26.0374 5.96247C26.457 6.38214 26.6668 6.89026 26.6668 7.48684V24.5125C26.6668 25.1091 26.457 25.6172 26.0374 26.0369C25.6177 26.4565 25.1096 26.6663 24.513 26.6663H7.48733ZM7.48733 25.333H24.513C24.7523 25.333 24.9489 25.2561 25.1027 25.1022C25.2566 24.9484 25.3335 24.7518 25.3335 24.5125V9.33301H6.66683V24.5125C6.66683 24.7518 6.74375 24.9484 6.8976 25.1022C7.05144 25.2561 7.24802 25.333 7.48733 25.333ZM9.3335 15.333V13.9997H22.6668V15.333H9.3335ZM9.3335 20.6663V19.333H17.3335V20.6663H9.3335Z"
+              fill="#1F1F1F"
+            />
+          </svg>
+        </IconContainer>
+      </NavLink>
+      <NavLink
+        to={`/mypage`}
+        className={({ isActive }) => (isActive ? 'active' : 'inactive')}
+      >
+        <IconContainer>
+          <svg
+            width="32"
+            height="32"
+            viewBox="0 0 32 32"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              d="M13.513 28L13.031 24.1436C12.6053 24.0154 12.1451 23.8137 11.6502 23.5385C11.1553 23.2632 10.7344 22.9684 10.3873 22.6539L6.83095 24.1667L4.34375 19.8333L7.41812 17.5154C7.37881 17.2743 7.34633 17.0261 7.32068 16.7705C7.29504 16.5149 7.28222 16.2667 7.28222 16.0256C7.28222 15.8017 7.29504 15.5662 7.32068 15.3192C7.34633 15.0722 7.37881 14.794 7.41812 14.4846L4.34375 12.1667L6.83095 7.8846L10.3617 9.3718C10.76 9.04018 11.1908 8.74103 11.654 8.47437C12.1173 8.2077 12.5677 8.00171 13.0053 7.8564L13.513 4H18.4874L18.9694 7.88203C19.4805 8.06152 19.9322 8.26751 20.3246 8.5C20.7169 8.73247 21.1207 9.02307 21.5361 9.3718L25.1694 7.8846L27.6566 12.1667L24.4797 14.5615C24.5532 14.8368 24.5942 15.0893 24.6028 15.3192C24.6113 15.5491 24.6156 15.7761 24.6156 16C24.6156 16.2068 24.607 16.4252 24.59 16.6551C24.5728 16.885 24.5335 17.1632 24.472 17.4897L27.5976 19.8333L25.1105 24.1667L21.5361 22.6282C21.1207 22.9769 20.7028 23.2761 20.2822 23.5256C19.8617 23.7752 19.4241 23.9727 18.9694 24.118L18.4874 28H13.513ZM14.6669 26.6667H17.2746L17.7669 23.0564C18.4387 22.8786 19.0433 22.6346 19.5809 22.3244C20.1185 22.0141 20.666 21.5889 21.2233 21.0487L24.5489 22.4667L25.8746 20.2L22.9566 18.0077C23.0677 17.6282 23.1412 17.2791 23.1771 16.9603C23.213 16.6415 23.2309 16.3214 23.2309 16C23.2309 15.6615 23.213 15.3414 23.1771 15.0397C23.1412 14.738 23.0677 14.406 22.9566 14.0436L25.9258 11.8L24.6002 9.53333L21.1976 10.959C20.7942 10.5162 20.2639 10.1013 19.6066 9.7141C18.9493 9.32692 18.3275 9.07009 17.7412 8.9436L17.3335 5.33333H14.6746L14.2592 8.91793C13.5874 9.06153 12.9698 9.29273 12.4066 9.61153C11.8433 9.93033 11.2831 10.3684 10.7258 10.9256L7.40018 9.53333L6.07455 11.8L8.96685 13.959C8.85574 14.2581 8.77796 14.5829 8.73352 14.9333C8.68907 15.2838 8.66685 15.6479 8.66685 16.0256C8.66685 16.3641 8.68907 16.7 8.73352 17.0333C8.77796 17.3667 8.84719 17.6915 8.94122 18.0077L6.07455 20.2L7.40018 22.4667L10.7002 21.0667C11.2232 21.5949 11.7664 22.0184 12.3297 22.3372C12.8929 22.656 13.5275 22.9043 14.2335 23.0821L14.6669 26.6667ZM15.9643 19.3333C16.8942 19.3333 17.6822 19.0103 18.3284 18.3641C18.9745 17.7179 19.2976 16.9299 19.2976 16C19.2976 15.0701 18.9745 14.2821 18.3284 13.6359C17.6822 12.9897 16.8942 12.6667 15.9643 12.6667C15.0292 12.6667 14.2399 12.9897 13.5963 13.6359C12.9527 14.2821 12.6309 15.0701 12.6309 16C12.6309 16.9299 12.9527 17.7179 13.5963 18.3641C14.2399 19.0103 15.0292 19.3333 15.9643 19.3333Z"
+              fill="#1F1F1F"
+            />
+          </svg>
+        </IconContainer>
+      </NavLink>
+    </Container>
+  );
+};
+
+export default TabBar;
